+++ src/App.js
@@ -8,7 +8,7 @@ import HomePage from "./pages/Home";
 import AccountPage from "./pages/Account";
 import AlbumPage from "./pages/Album";
 import NotFound from "./pages/NotFound";
-import Spotify from "./database/spotify";
+import Spotify from "./util/spotify";
 
 import { signInWithEmailAndPassword } from "firebase/auth";
 import { auth, db } from "./database/firebase";
