+++ src/pages/Home/index.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Header from '../../components/Header/Header';
 import Footer from '../../components/Footer/Footer';
 import { useNavigation, Link } from 'react-router-dom';
-import spotify from "../../database/spotify";
+import spotify from "../../util/spotify";
 
 import "./home.scss";
